moving class related to json serialization in the json formatter project
