+++ src/App.Metrics.Formatters.Json/ApdexMetric.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-using App.Metrics.Core.Abstractions;
-
-namespace App.Metrics.Apdex
+namespace App.Metrics.Formatters.Json
 {
     public sealed class ApdexMetric : MetricBase
     {
