+++ src/App.Metrics.Formatters.Json/CounterMetric.cs
@@ -3,10 +3,8 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using App.Metrics.Core;
-using App.Metrics.Core.Abstractions;
 
-namespace App.Metrics.Counter
+namespace App.Metrics.Formatters.Json
 {
     /// <summary>
     ///     <para>
