+++ src/App.Metrics.Formatters.Json/GaugeMetric.cs
@@ -1,11 +1,10 @@
 ﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-using App.Metrics.Core;
 using App.Metrics.Core.Abstractions;
 using App.Metrics.Tagging;
 
-namespace App.Metrics.Gauge
+namespace App.Metrics.Formatters.Json
 {
     /// <summary>
     ///     <para>
