+++ src/App.Metrics.Formatters.Json/HistogramMetric.cs
@@ -2,10 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using App.Metrics.Abstractions.ReservoirSampling;
-using App.Metrics.Core;
-using App.Metrics.Core.Abstractions;
 
-namespace App.Metrics.Histogram
+namespace App.Metrics.Formatters.Json
 {
     /// <summary>
     ///     <para>
