+++ src/App.Metrics.Formatters.Json/MetricBase.cs
@@ -4,7 +4,7 @@
 using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace App.Metrics.Core.Abstractions
+namespace App.Metrics.Formatters.Json
 {
     public abstract class MetricBase
     {
