+++ src/App.Metrics.Formatters.Json/MetricsContext.cs
@@ -3,15 +3,9 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using App.Metrics.Apdex;
-using App.Metrics.Counter;
-using App.Metrics.Gauge;
-using App.Metrics.Histogram;
-using App.Metrics.Meter;
-using App.Metrics.Timer;
 using Newtonsoft.Json;
 
-namespace App.Metrics.Core
+namespace App.Metrics.Formatters.Json
 {
     /// <summary>
     ///     MetricsContext's are a way of grouping metrics withing a context, for example we can record all database related
