+++ src/App.Metrics.Formatters.Json/Serialization/HealthStatusSerializer.cs
@@ -25,7 +25,7 @@ namespace App.Metrics.Formatters.Json.Serialization
 
         public HealthStatusSerializer(JsonSerializerSettings serializerSettings) { _settings = serializerSettings; }
 
-        public virtual T Deserialize<T>(string json) { return JsonConvert.DeserializeObject<T>(json, _settings); }
+        public virtual T Deserialize<T>(string value) { return JsonConvert.DeserializeObject<T>(value, _settings); }
 
         public virtual string Serialize<T>(T value) { return JsonConvert.SerializeObject(value, _settings); }
     }
