+++ src/App.Metrics/Histogram/Extensions/HistogramExtensions.cs
@@ -1,13 +1,10 @@
 ﻿// Copyright (c) Allan Hardy. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-using System.Collections.Generic;
-using System.Linq;
 using App.Metrics.Abstractions.MetricTypes;
 using App.Metrics.Core;
 using App.Metrics.Core.Abstractions;
 using App.Metrics.Histogram.Abstractions;
-using App.Metrics.Tagging;
 
 // ReSharper disable CheckNamespace
 namespace App.Metrics.Histogram
@@ -42,68 +39,6 @@ namespace App.Metrics.Histogram
         {
             var implementation = metric as IHistogramMetric;
             return implementation != null ? implementation.Value : EmptyHistogram;
-        }
-
-        public static IEnumerable<HistogramMetric> ToMetric(this IEnumerable<HistogramValueSource> source) { return source.Select(ToMetric); }
-
-        public static HistogramMetric ToMetric(this HistogramValueSource source)
-        {
-            return new HistogramMetric
-                   {
-                       Name = source.Name,
-                       Group = source.Group,
-                       Count = source.Value.Count,
-                       Unit = source.Unit.Name,
-                       LastUserValue = source.Value.LastUserValue,
-                       LastValue = source.Value.LastValue,
-                       Max = source.Value.Max,
-                       MaxUserValue = source.Value.MaxUserValue,
-                       Mean = source.Value.Mean,
-                       Median = source.Value.Median,
-                       Min = source.Value.Min,
-                       MinUserValue = source.Value.MinUserValue,
-                       Percentile75 = source.Value.Percentile75,
-                       Percentile95 = source.Value.Percentile95,
-                       Percentile98 = source.Value.Percentile98,
-                       Percentile99 = source.Value.Percentile99,
-                       Percentile999 = source.Value.Percentile999,
-                       SampleSize = source.Value.SampleSize,
-                       StdDev = source.Value.StdDev,
-                       Tags = source.Tags.ToDictionary()
-                   };
-        }
-
-        public static HistogramValueSource ToMetricValueSource(this HistogramMetric source)
-        {
-            var histogramValue = new HistogramValue(
-                source.Count,
-                source.LastValue,
-                source.LastUserValue,
-                source.Max,
-                source.MaxUserValue,
-                source.Mean,
-                source.Min,
-                source.MinUserValue,
-                source.StdDev,
-                source.Median,
-                source.Percentile75,
-                source.Percentile95,
-                source.Percentile98,
-                source.Percentile99,
-                source.Percentile999,
-                source.SampleSize);
-
-            return new HistogramValueSource(
-                source.Name,
-                source.Group,
-                ConstantValue.Provider(histogramValue),
-                source.Unit,
-                source.Tags.FromDictionary());
-        }
-
-        public static IEnumerable<HistogramValueSource> ToMetricValueSource(this IEnumerable<HistogramMetric> source)
-        {
-            return source.Select(x => x.ToMetricValueSource());
         }
     }
 }
