+++ src/App.Metrics/Serialization/NoOpHealthStatusSerializer.cs
@@ -9,7 +9,7 @@ namespace App.Metrics.Serialization
     [AppMetricsExcludeFromCodeCoverage]
     public sealed class NoOpHealthStatusSerializer : IHealthStatusSerializer
     {
-        public T Deserialize<T>(string json) { return default(T); }
+        public T Deserialize<T>(string value) { return default(T); }
 
         public string Serialize<T>(T value) { return string.Empty; }
     }
